rustc_feature

Module unstable

List of the unstable feature gates.

Macros

declare_features 🔒

status_to_enum 🔒

Structs

EnabledLangFeature

Information about an enabled language feature.

EnabledLibFeature

Information about an enabled library feature.

Features

A set of features to be used by later passes.

Enums

FeatureStatus 🔒

Constants

INCOMPATIBLE_FEATURES

Some features are not allowed to be used together at the same time, if the two are present, produce an error.

Statics

UNSTABLE_LANG_FEATURES

Unstable language features that are being implemented or being considered for acceptance (stabilization) or removal.