Crate rustc_feature

Feature gates

This crate declares the set of past and present unstable features in the compiler. Feature gate checking itself is done in rustc_ast_passes/src/feature_gate.rs at the moment.

Features are enabled in programs via the crate-level attributes of #![feature(...)] with a comma-separated list of features.

For the purpose of future feature-tracking, once a feature gate is added, even if it is stabilized or removed, do not remove it. Instead, move the symbol to the accepted or removed modules respectively.

Modules

accepted 🔒

List of the accepted feature gates.

active 🔒

List of the active feature gates.

builtin_attrs 🔒

Built-in attributes and cfg flag gating.

removed 🔒

List of the removed feature gates.

Structs

AttributeTemplate

A template that the attribute input must match. Only top-level shape (#[attr] vs #[attr(...)] vs #[attr = ...]) is considered now.

BuiltinAttribute

Feature

Features

A set of features to be used by later passes.

Enums

AttributeDuplicates

How to handle multiple duplicate attributes on the same item.

AttributeGate

AttributeType

GateIssue

Stability

State

UnstableFeatures

Constants

ACCEPTED_FEATURES

Those language feature has since been Accepted (it was once Active)

ACTIVE_FEATURES

Represents active features that are currently being implemented or currently being considered for addition/removal.

BUILTIN_ATTRIBUTES

Attributes that have a special meaning to rustc or rustdoc.

INCOMPATIBLE_FEATURES

Some features are not allowed to be used together at the same time, if the two are present, produce an error.

REMOVED_FEATURES

Represents unstable features which have since been removed (it was once Active)

STABLE_REMOVED_FEATURES

Represents stable features which have since been removed (it was once Accepted)

Statics

BUILTIN_ATTRIBUTE_MAP

Functions

deprecated_attributes

find_feature_issue

find_gated_cfg

Find a gated cfg determined by the predicate which is given the cfg’s name.

find_lang_feature_issue 🔒

is_builtin_attr_name

is_builtin_only_local

Whether this builtin attribute is only used in the local crate. If so, it is not encoded in the crate metadata.

is_valid_for_get_attr

to_nonzero 🔒

Type Aliases

GatedCfg