Module rustc_feature::unstable

List of the unstable feature gates.

Macros

declare_features 🔒

status_to_enum 🔒

Structs

Features

A set of features to be used by later passes.

UnstableFeature

Enums

FeatureStatus 🔒

Constants

INCOMPATIBLE_FEATURES

Some features are not allowed to be used together at the same time, if the two are present, produce an error.

NUM_FEATURES 🔒

UNSTABLE_FEATURES

Unstable language features that are being implemented or being considered for acceptance (stabilization) or removal.